@@ -1,27 +1,29 @@
 package uk.org.landeg.kalah.api.transformer;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import uk.org.landeg.kalah.api.model.CreateGameResponseModel;
 import uk.org.landeg.kalah.components.KalahGameState;
 import uk.org.landeg.kalah.exception.KalahException;
 
-public class CreateGameResponseTransformerTest {
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class CreateGameResponseTransformerTest {
 	private CreateGameResponseTransformer transformer = new CreateGameResponseTransformer();
 	
 	@Test
-	public void assertTransformToRestAsExpected() {
+	void assertTransformToRestAsExpected() {
 		KalahGameState state = new KalahGameState();
 		state.setGameId(123456L);
 		state.setUrl("http://dont-care-where");
 		final CreateGameResponseModel response = transformer.toRest(state);
-		Assert.assertEquals(state.getGameId().toString(), response.getId());
-		Assert.assertEquals(state.getUrl().toString(), response.getUrl());
+		assertThat(state.getGameId()).hasToString(response.getId());
+		assertThat(state.getUrl()).hasToString(response.getUrl());
 	}
 	
-	@Test(expected = KalahException.class)
-	public void assertExceptionOnNullEntity() {
-		transformer.toRest(null);
+	@Test
+	void assertExceptionOnNullEntity() {
+		assertThrows(KalahException.class, () -> transformer.toRest(null));
 	}
 }
