@@ -1,27 +1,29 @@
 package uk.org.landeg.kalah.game.action;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.test.context.junit4.SpringRunner;
-
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.org.landeg.kalah.CommonTestConfiguration;
 import uk.org.landeg.kalah.components.KalahGameState;
+import uk.org.landeg.kalah.game.KalahGameBoard;
 import uk.org.landeg.kalah.game.KalahGameEngine;
 
-@RunWith(SpringRunner.class)
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+@ExtendWith(SpringExtension.class)
 @Import(CommonTestConfiguration.class)
-public class CaptureKalahActionTest {
+class CaptureKalahActionTest {
 	@TestConfiguration
 	static class Config {
 		@Bean
-		KalahAction captureAction() {
-			return new CaptureKalahAction();
+		@Autowired
+		KalahAction captureAction(KalahGameBoard gameBoard) {
+			return new CaptureKalahAction(gameBoard);
 		}
 	}
 
@@ -37,31 +39,31 @@ public class CaptureKalahActionTest {
 	 * expected capture 1 + 2 stones.
 	 */
 	@Test
-	public void assertRuleAppliesWhenExpected() {
+	void assertRuleAppliesWhenExpected() {
 		final KalahGameState game = new KalahGameState();
 		service.initialiseGame(game, 3);
 		game.setRecentPit(4);
 		game.getPits().put(1, 3);
 		game.getPits().put(10, 2);
 		game.getPits().put(4, 1);
-		Assert.assertTrue(action.applies(game));
+		assertTrue(action.applies(game));
 		action.processAction(game);
-		Assert.assertEquals(0, game.getPits().get(4).intValue());
-		Assert.assertEquals(0, game.getPits().get(10).intValue());
-		Assert.assertEquals(3, game.getPits().get(7).intValue());
+		assertThat(0).isEqualTo(game.getPits().get(4).intValue());
+		assertThat(0).isEqualTo(game.getPits().get(10).intValue());
+		assertThat(3).isEqualTo(game.getPits().get(7).intValue());
 	}
 
 	/**
 	 * identical scenario to above, but last pit was not empty 
 	 */
 	@Test
-	public void assertRuleRejectedWhenExpected() {
+	void assertRuleRejectedWhenExpected() {
 		final KalahGameState game = new KalahGameState();
 		service.initialiseGame(game, 3);
 		game.setRecentPit(4);
 		game.getPits().put(1, 3);
 		game.getPits().put(10, 2);
 		game.getPits().put(4, 2);
-		Assert.assertFalse(action.applies(game));
+		assertFalse(action.applies(game));
 	}
 }
