@@ -1,19 +1,18 @@
 package uk.org.landeg.kalah.game.action;
 
-import static org.junit.Assert.assertEquals;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.mockito.Mockito.when;
 
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import uk.org.landeg.kalah.CommonTestConfiguration;
 import uk.org.landeg.kalah.Constants.Player;
@@ -22,14 +21,15 @@ import uk.org.landeg.kalah.game.KalahGameBoard;
 import uk.org.landeg.kalah.game.KalahGameEngine;
 import uk.org.landeg.kalah.game.KalahPitDecorator;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @Import(CommonTestConfiguration.class)
-public class FinaliseGameKalahActionTest {
+class FinaliseGameKalahActionTest {
 	@TestConfiguration
 	static class Config {
 		@Bean
-		KalahAction finaliseGameKalahAction() {
-			return new FinaliseGameKalahAction();
+		@Autowired
+		KalahAction finaliseGameKalahAction(KalahGameBoard gameBoard) {
+			return new FinaliseGameKalahAction(gameBoard);
 		}
 	}
 
@@ -48,31 +48,31 @@ public class FinaliseGameKalahActionTest {
 	@Mock
 	KalahPitDecorator pits;
 
-	@Before
-	public void setup() {
+	@BeforeEach
+	void setup() {
 		when(game.getPits()).thenReturn(pits);
 	}
 
 	@Test
-	public void assertApplicableForZeroNorthStones() {
+	void assertApplicableForZeroNorthStones() {
 		when(pits.get(Mockito.any())).thenAnswer(invocation -> {
 			Integer pitId = invocation.getArgument(0);
 			return gameBoard.getPlayerPits().get(Player.NORTH).contains(pitId) ? 0 : 1;
 		});
-		Assert.assertTrue(finaliseAction.applies(game));
+		assertThat(finaliseAction.applies(game)).isTrue();
 	}
 
 	@Test
-	public void assertApplicableForZeroSouthStones() {
+	void assertApplicableForZeroSouthStones() {
 		when(pits.get(Mockito.any())).thenAnswer(invocation -> {
 			Integer pitId = invocation.getArgument(0);
 			return gameBoard.getPlayerPits().get(Player.SOUTH).contains(pitId) ? 0 : 1;
 		});
-		Assert.assertTrue(finaliseAction.applies(game));
+		assertThat(finaliseAction.applies(game)).isTrue();
 	}
 
 	@Test
-	public void assertNotApplicableWhenStonesRemain() {
+	void assertNotApplicableWhenStonesRemain() {
 		when(pits.get(Mockito.any())).thenAnswer(invocation -> {
 			Integer pitId = invocation.getArgument(0);
 			int southPit = gameBoard.getPlayerPits().get(Player.SOUTH).get(0);
@@ -80,12 +80,12 @@ public class FinaliseGameKalahActionTest {
 			// simulate one stone of each side of the board. 
 			return (pitId == southPit || pitId == northPit) ? 1 : 0;
 		});
-		Assert.assertFalse(finaliseAction.applies(game));
+		assertThat(finaliseAction.applies(game)).isFalse();
 	}
 
 
 	@Test
-	public void processingSuccessful() {
+	void processingSuccessful() {
 		final KalahGameState game = new KalahGameState();
 		gameService.initialiseGame(game);
 		// put 1 stone in each south pit, 2 in each north pit
@@ -98,7 +98,10 @@ public class FinaliseGameKalahActionTest {
 		
 		finaliseAction.processAction(game);
 
-		assertEquals(expectedNorthStones, pits.get(gameBoard.getPlayerKalah().get((Player.NORTH))).intValue());
-		assertEquals(expectedSouthStones, pits.get(gameBoard.getPlayerKalah().get((Player.SOUTH))).intValue());
+		var northKalah = gameBoard.getPlayerKalah().get((Player.NORTH));
+		var southKalah = gameBoard.getPlayerKalah().get((Player.SOUTH));
+		assertThat(pits)
+				.containsEntry(northKalah, expectedNorthStones)
+				.containsEntry(southKalah, expectedSouthStones);
 	}
 }
