@@ -1,18 +1,19 @@
 package uk.org.landeg.kalah.game.api;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -28,16 +29,16 @@ import uk.org.landeg.kalah.service.KalahService;
  * @author Andrew Landeg
  *
  */
-@RunWith(MockitoJUnitRunner.class)
-public class GameApiTest {
+@ExtendWith(MockitoExtension.class)
+class GameApiTest {
 	@Mock
 	KalahService kalahWebService;
 
 	@InjectMocks
-	GameApi api = new GameApi();
+	GameApi api;
 
 	@Test
-	public void assertCreateGameBehavior() {
+	void assertCreateGameBehavior() {
 		int stones = KalahGameBoardStandard.DEFAULT_INIT_STONES;
 		final CreateGameResponseModel modelResponse = new CreateGameResponseModel();
 		when(kalahWebService.createGame(stones)).
@@ -45,12 +46,12 @@ public class GameApiTest {
 		final ResponseEntity<CreateGameResponseModel> response = api.createGame(Optional.empty());
 		// assert the service is correctly invoked, and that something sensible is returned.
 		verify(kalahWebService, times(1)).createGame(stones);
-		Assert.assertEquals(HttpStatus.CREATED, response.getStatusCode());
-		Assert.assertSame(modelResponse, response.getBody());
+		assertThat(HttpStatus.CREATED).isEqualTo(response.getStatusCode());
+		assertSame(modelResponse, response.getBody());
 	}
 
 	@Test
-	public void assertMakeMoveBehavior() {
+	void assertMakeMoveBehavior() {
 		long gameId = 123456;
 		int pitId = 1;
 		final PerformMoveResponseModel repsonseModel = new PerformMoveResponseModel(); 
@@ -58,6 +59,6 @@ public class GameApiTest {
 			.thenReturn(repsonseModel);
 		final PerformMoveResponseModel actualResponse = api.makeMove(gameId, pitId);
 		verify(kalahWebService, times(1)).makeMove(gameId, pitId);
-		Assert.assertSame(repsonseModel, actualResponse);
+		assertSame(repsonseModel, actualResponse);
 	}
 }
