@@ -1,23 +1,23 @@
-package uk.org.landeg.kalah;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@Configuration
-public class SwaggerConfiguration {
-	@Bean
-	public Docket swaggerDocket() {
-        return new Docket(DocumentationType.SWAGGER_2)  
-            .select()                                  
-            .apis(RequestHandlerSelectors.any())              
-            .paths(PathSelectors.any())                          
-            .build();                                           
-	}
-}
+//package uk.org.landeg.kalah;
+//
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//
+//import springfox.documentation.builders.PathSelectors;
+//import springfox.documentation.builders.RequestHandlerSelectors;
+//import springfox.documentation.spi.DocumentationType;
+//import springfox.documentation.spring.web.plugins.Docket;
+//import springfox.documentation.swagger2.annotations.EnableSwagger2;
+//
+//@EnableSwagger2
+//@Configuration
+//public class SwaggerConfiguration {
+//	@Bean
+//	public Docket swaggerDocket() {
+//        return new Docket(DocumentationType.SWAGGER_2)
+//            .select()
+//            .apis(RequestHandlerSelectors.any())
+//            .paths(PathSelectors.any())
+//            .build();
+//	}
+//}
