@@ -1,5 +1,6 @@
 package uk.org.landeg.kalah.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -8,6 +9,7 @@ import uk.org.landeg.kalah.api.model.CreateGameResponseModel;
 import uk.org.landeg.kalah.api.model.PerformMoveResponseModel;
 import uk.org.landeg.kalah.api.transformer.RestEntityTransformer;
 import uk.org.landeg.kalah.components.KalahGameState;
+import uk.org.landeg.kalah.exception.KalahClientException;
 import uk.org.landeg.kalah.exception.KalahGameNotFoundException;
 import uk.org.landeg.kalah.game.KalahGameBoard;
 import uk.org.landeg.kalah.game.KalahGameEngine;
@@ -15,26 +17,21 @@ import uk.org.landeg.kalah.persistence.PersistenceService;
 import uk.org.landeg.kalah.resolver.RequestUriResolver;
 
 @Service
+@RequiredArgsConstructor
 public class KalahService {
-	@Autowired
-	KalahGameEngine gameService;
+	private final KalahGameEngine gameService;
 
-	@Autowired
-	KalahGameBoard gameBoard;
+	private final KalahGameBoard gameBoard;
 
-	@Autowired
-	PersistenceService persistenceService;
+	private final PersistenceService persistenceService;
 
-	@Autowired
-	RequestUriResolver uriResolver;
+	private final RequestUriResolver uriResolver;
 
-	@Autowired
 	@RestTransformer(restEntity = PerformMoveResponseModel.class)
-	RestEntityTransformer<PerformMoveResponseModel, KalahGameState> performMoveTransformer;
+	private final RestEntityTransformer<PerformMoveResponseModel, KalahGameState> performMoveTransformer;
 
-	@Autowired
 	@RestTransformer(restEntity=CreateGameResponseModel.class)
-	RestEntityTransformer<CreateGameResponseModel, KalahGameState> createGameTransformer;
+	private final RestEntityTransformer<CreateGameResponseModel, KalahGameState> createGameTransformer;
 
 	/**
 	 * Create a new Kalah game instance.
@@ -86,6 +83,7 @@ public class KalahService {
 	}
 
 	private KalahGameState retrieveGameInternal(final Long gameId) {
-		return persistenceService.findById(gameId).orElseThrow(() -> new KalahGameNotFoundException());
+		return persistenceService.findById(gameId)
+				.orElseThrow(KalahGameNotFoundException::new);
 	}
 }
